requirements.txt
certifi==2020.6.20
charset-normalizer==2.0.11
click==8.0.3
EasyProcess==1.1
entrypoint2==1.0
Flask==2.0.2
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.2
pandas==1.4.0
patool==1.12
python-dateutil==2.8.2
pytz==2021.3
pyunpack==0.2.2
rarfile==4.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
gunicorn
